[RFT] change model data type

+++ datasource/src/main/java/net/hdhuu/datasource/model/PostDataEntity.kt
@@ -3,4 +3,4 @@ package net.hdhuu.datasource.model
 /**
  * Created by Huu Hoang on 2/16/19.
  */
-data class PostDataEntity(val id: Long, val content: String, val createAt: Double)
+data class PostDataEntity(val id: Int, val content: String, val createAt: Double)
